# contracts/openapi/learning.openapi.yaml
openapi: 3.0.3
info:
  title: Learning Service (Impilo Fundo) — Internal API
  description: >
    Internal API surface for **Impilo Fundo**, the **native learning
    management, certification, in-service training, pre-service training
    and CPD support capability of Impilo vNext**. Fundo is designed to
    operate as a complete modern LMS at national scale — its own
    catalogue, paths, enrolment, progress, assessments, certificates,
    learning records and analytics — without depending on any external
    LMS to function.


    External LMS platforms (Moodle, Open edX, Canvas, Chamilo, …) may be
    connected through **optional adapters** where useful, but Fundo must
    not depend on any external LMS. The existing Moodle orchestration
    endpoints documented below are one such optional adapter, preserved
    verbatim for backwards compatibility — not the Fundo runtime.


    This contract is **descriptive of the current implementation only** —
    it documents the endpoints that exist in `services/learning-service`
    today and does not assert new behaviour, new fields, or future
    surfaces. Endpoints under `/internal/v1/learning/integrations/**`
    and `/internal/v1/learning/orchestration/**` are server-to-server
    only and use shared internal keys (see `securitySchemes`), not
    end-user JWT.


    CPD authority remains with `varapi-service`. The Impilo Fundo brand
    is the product/UX name; `learning-service` is its technical home.
  version: 0.1.0
...
servers:
...
paths:
...
  /internal/v1/learning/resource-opened:
    post:
      tags: [Internal]
      summary: Record that a user opened a learning resource (audit + progress)
      operationId: recordResourceOpened
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceOpenedRequest' }
      responses:
        '200':
          description: Event recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Status' }
...
  /internal/v1/learning/v11/resource-opened-acknowledgements:
    post:
      tags: [InternalV11]
      summary: Acknowledge that a learning resource was opened — native v1.1 write surface (Phase 4A)
      description: >
        Additive **v1.1-compliant** write endpoint for Impilo Fundo. Delegates
        to the same native facade method as the legacy
        `POST /internal/v1/learning/resource-opened`
        (`LearningPlatformFacade.recordResourceOpened`) and never invokes
        Moodle or any external LMS. Enforces the platform v1.1 contract
        (companion header enforcement, error envelope, client-timeout
        enforcement, `Idempotency-Key` requirement) via the shared
        tech-companion filter chain. The legacy endpoint remains unchanged
        and continues to serve BFF traffic.


        **Tolerant inputs.** When `resourceId` is missing or not a valid UUID,
        or when the `X-Tenant-ID` header is not a valid UUID, the endpoint
        returns a side-effect-free
        `{"data": {"status": "acknowledged", "noOp": true, "reason": "..."}}`
        envelope with HTTP 200 rather than 4xx. This keeps the v1.1 envelope
        contract honest (2xx for "all headers + idempotency key present"
        callers) without leaking domain-specific validation through the
        platform error envelope. Production callers supplying a valid UUID
        tenant and `resourceId` receive `{"data": {"status": "recorded",
        "resourceId": "..."}}`.
      operationId: acknowledgeResourceOpenedV11
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                resourceId:
                  type: string
                  format: uuid
                actorId: { type: string }
                actorType: { type: string }
                workflowContext:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Acknowledgement recorded or accepted as no-op (v1.1 envelope)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Map' }
        '400':
          description: Required v1.1 header missing or `Idempotency-Key` absent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Same `Idempotency-Key` reused with a different body (`IDENTITY_CONFLICT`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '504':
          description: Client timeout exceeded (`X-Client-Timeout-MS` reached)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────────────────────────────────────────────────
  # Phase 5 — Native Impilo Fundo LMS surface
  #
  # All endpoints below run on the native LMS schema (V006 / V007) and
  # never invoke Moodle, Open edX, Canvas, Chamilo, SCORM runtime, xAPI
  # runtime, or any other external LMS. They are governed by the same
  # v1.1 platform contract as `subject-completions` and
  # `resource-opened-acknowledgements` above (companion header
  # enforcement, error envelope, idempotency, client-timeout
  # enforcement).
  # ───────────────────────────────────────────────────────────────────────

  # ───────────────────────────────────────────────────────────────────────
  # Phase 6 — Native Fundo authoring + delivery layer additions
  #
  # Authoring write endpoints (Phase 6A):
  #   POST   /v11/catalog
  #   PUT    /v11/catalog/{courseId}
  #   POST   /v11/courses/{courseId}/modules
  #   PUT    /v11/modules/{moduleId}
  #   POST   /v11/modules/{moduleId}/lessons
  #   PUT    /v11/lessons/{lessonId}
  #   POST   /v11/pathways
  #   PUT    /v11/pathways/{pathwayId}
  #   POST   /v11/pathways/{pathwayId}/items
  #   POST   /v11/assessments
  #   PUT    /v11/assessments/{assessmentId}
  #   POST   /v11/assessments/{assessmentId}/questions
  #
  # Cohort report (Phase 6C):
  #   GET    /v11/reports/cohort-completions
  #
  # All write endpoints emit native v1.1 events where applicable; the
  # `DRAFT → PUBLISHED` course transition emits
  # `impilo.learning.course.published.v1` (constant reserved in Phase 5D,
  # finally emitted in Phase 6A). No Moodle / external LMS dependency
  # is introduced.
  # ───────────────────────────────────────────────────────────────────────
...
  /internal/v1/learning/v11/assessments/{assessmentId}:
    get:
      tags: [FundoNativeLMS]
      summary: Get assessment with its questions (correct answers omitted)
      operationId: getFundoAssessment
      parameters:
        - { name: assessmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Assessment detail
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    put:
      tags: [FundoNativeLMS]
      summary: Update a native assessment (Phase 6A)
      operationId: updateFundoAssessment
      parameters:
        - { name: assessmentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FundoAssessmentUpsert' }
      responses:
        '200':
          description: Assessment updated
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  /internal/v1/learning/v11/assessments/{assessmentId}/attempts:
    post:
      tags: [FundoNativeLMS]
      summary: Submit a native assessment attempt
      description: >
        Records an attempt with simple objective scoring for
        `MULTIPLE_CHOICE` and `TRUE_FALSE` question types only. Attempts
        that include `SHORT_ANSWER`, `MATCHING` or `CASE_PROMPT` questions
        are accepted with `score=null` and `passed=null` (pending manual
        review). Emits `impilo.learning.assessment.attempt.submitted.v1`.
      operationId: submitFundoAssessmentAttempt
      parameters:
        - { name: assessmentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectType, subjectId]
              properties:
                enrolmentId:  { type: string, format: uuid, nullable: true }
                subjectType:  { type: string }
                subjectId:    { type: string }
                answers:
                  type: object
                  additionalProperties: true
                  description: 'Map of questionId → answer.'
      responses:
        '200':
          description: Attempt recorded
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Assessment not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: Maximum attempts reached
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Used by the default security chain when `learning.security.oauth2-enabled=true`.
        Compose dev environments run with this disabled (see
        `ops/runtime/docker-compose.operations.yml`); Helm defaults to true.
    integrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Internal-Key
      description: >
        Shared internal key for trusted ingress from `varapi-service` after a
        Fundo webhook creates a governed CPD candidate.
    orchestrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Orchestration-Key
      description: >
        Shared internal key for server-to-server orchestration jobs (Tshepo /
        Varapi schedulers, Moodle Web Service relays).
  schemas:
...
    ApiEnvelope_Map:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
      required: [data]
...
    ApiEnvelope_Status:
      type: object
      properties:
        data:
          type: object
          properties:
            status: { type: string }
          required: [status]
      required: [data]
...
    Error:
      type: object
      properties:
        error: { type: string }

    # ── Request bodies ────────────────────────────────────────────────────
    ResourceOpenedRequest:
      type: object
      properties:
        resourceId: { type: string, format: uuid }
        workflowContext:
          type: object
          additionalProperties: true
      required: [resourceId]
...
    FundoAssessmentUpsert:
      type: object
      properties:
        courseId: { type: string, format: uuid, nullable: true, description: "Required on create" }
        moduleId: { type: string, format: uuid, nullable: true }
        title: { type: string, nullable: true, description: "Required on create" }
        description: { type: string, nullable: true }
        assessmentType: { type: string, nullable: true, description: "Required on create. QUIZ | PRACTICAL_CHECKLIST | CASE_REVIEW" }
        passMark: { type: integer, nullable: true }
        maxAttempts: { type: integer, nullable: true }
        status: { type: string, nullable: true }
...